refactor(player): split nav widgets out of the app shell
