+++ player/lib/presentation/widgets/nav/back_to_mydia_button.dart
@@ -1,0 +1,59 @@
+import 'package:flutter/material.dart';
+
+import '../../../core/config/web_config.dart';
+import '../../../core/theme/colors.dart';
+
+/// Button to navigate back to the main Mydia app.
+/// Shown when the player is running in embed mode.
+class BackToMydiaButton extends StatefulWidget {
+  const BackToMydiaButton({super.key});
+
+  @override
+  State<BackToMydiaButton> createState() => _BackToMydiaButtonState();
+}
+
+class _BackToMydiaButtonState extends State<BackToMydiaButton> {
+  bool _isHovered = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return MouseRegion(
+      onEnter: (_) => setState(() => _isHovered = true),
+      onExit: (_) => setState(() => _isHovered = false),
+      child: GestureDetector(
+        onTap: navigateToMydiaApp,
+        child: AnimatedContainer(
+          duration: const Duration(milliseconds: 200),
+          curve: Curves.easeOutCubic,
+          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+          decoration: BoxDecoration(
+            color: _isHovered
+                ? AppColors.surfaceVariant.withValues(alpha: 0.5)
+                : Colors.transparent,
+            borderRadius: BorderRadius.circular(12),
+          ),
+          child: Row(
+            children: [
+              Icon(
+                Icons.arrow_back_rounded,
+                size: 20,
+                color: _isHovered ? AppColors.primary : AppColors.textSecondary,
+              ),
+              const SizedBox(width: 12),
+              Text(
+                'Back to Mydia',
+                style: TextStyle(
+                  fontSize: 14,
+                  fontWeight: FontWeight.w500,
+                  color: _isHovered
+                      ? AppColors.textPrimary
+                      : AppColors.textSecondary,
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
